move around query builder result panel
