+++ .changeset/five-tomatoes-tease.md
@@ -1,0 +1,5 @@
+---
+'@finos/legend-application-studio': patch
+'@finos/legend-query-builder': patch
+'@finos/legend-lego': patch
+---
